refactor: :recycle: refactor DBSession class, move to app.db.session
