+++ app/api/api_v1/dependencies/__init__.py
@@ -5,7 +5,6 @@ from app.api.api_v1.dependencies.block_name_resolver import (
 )
 from app.api.api_v1.dependencies.block_search_params import BlockSearchParameters
 from app.api.api_v1.dependencies.char_search_params import CharacterSearchParameters
-from app.api.api_v1.dependencies.db_session import DBSession, get_session
 from app.api.api_v1.dependencies.filter_params import FilterParameters
 from app.api.api_v1.dependencies.list_params import ListParameters, ListParametersDecimal
 from app.api.api_v1.dependencies.plane_abbrev_resolver import UnicodePlaneResolver
