+++ app/api/api_v1/endpoints/characters.py
@@ -5,11 +5,9 @@ from fastapi import APIRouter, Depends, HTTPException, Path, Query, status
 import app.db.models as db
 from app.api.api_v1.dependencies import (
     CharacterSearchParameters,
-    DBSession,
     FilterParameters,
     ListParameters,
     UnicodeBlockQueryParamResolver,
-    get_session,
 )
 from app.api.api_v1.dependencies.filter_param_matcher import filter_param_matcher
 from app.api.api_v1.endpoints.util import get_character_details
@@ -17,6 +15,7 @@ from app.api.api_v1.pagination import paginate_search_results
 from app.config.api_settings import get_settings
 from app.data.cache import cached_data
 from app.data.encoding import get_codepoint_string
+from app.db.session import DBSession, get_session
 from app.docs.dependencies.custom_parameters import (
     UNICODE_CHAR_STRING_DESCRIPTION,
     VERBOSE_DESCRIPTION,
