--- app/db/procs/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from app.db.procs.get_char_details import get_character_properties
